# docs/openapi/v0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SUAS API",
    "version": "0.2.0",
    "description": "Machine-readable contract for the implemented `/api/v0` surface. Paths document what the runtime registers today; draft APIS.md Plane A paths that are not yet implemented are intentionally absent."
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the SUAS process"
    }
  ],
...
  "paths": {
...
    "/api/v0/trusted-contacts": {
      "get": {
        "tags": ["consent"],
        "summary": "List the veteran trusted circle",
        "operationId": "listTrustedContacts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trusted contacts without invite channel literals"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": ["consent"],
        "summary": "Invite a trusted contact",
        "operationId": "inviteTrustedContact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Invite recorded"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session credential from challenge verify"
      }
    },
...
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Canonical API.md §6 error body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        }
      }
    }
  }
}